Remove LangChain callback handler and related files, including utility functions and documentation, to streamline the AgentOps SDK. Updated semconv, excluded model provider info
