+++ agentops/integration/callbacks/langchain/__init__.py
@@ -4,7 +4,7 @@ LangChain integration for AgentOps.
 This module provides the AgentOps LangChain integration, including callbacks and utilities.
 """
 
-from agentops.sdk.callbacks.langchain.callback import (
+from agentops.integration.callbacks.langchain.callback import (
     LangchainCallbackHandler,
     AsyncLangchainCallbackHandler,
 )
